# BankService/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Payment Gateway",
    "version": "v1"
  },
  "paths": {
    "/api/valid": {
      "post": {
        "tags": [
          "Payment"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/CardDetails"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardDetails"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CardDetails"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CardDetails"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CheckCardResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckCardResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckCardResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/process": {
      "post": {
        "tags": [
          "Payment"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentDetails"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentDetails"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentDetails"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentDetails"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessPaymentResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessPaymentResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessPaymentResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/find/{id}": {
      "get": {
        "tags": [
          "Payment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "default": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRequestLog"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRequestLog"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRequestLog"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CardDetails": {
        "required": [
          "cardholderName",
          "cardNumber",
          "csc",
          "expires"
        ],
        "type": "object",
        "properties": {
          "cardholderName": {
            "type": "string"
          },
          "cardNumber": {
            "maxLength": 16,
            "minLength": 16,
            "pattern": "[0-9]+",
            "type": "string"
          },
          "expires": {
            "type": "string",
            "format": "date-time"
          },
          "validFrom": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "csc": {
            "maxLength": 3,
            "minLength": 3,
            "pattern": "[0-9]+",
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "CheckCardResult": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "TransactionDetails": {
        "required": [
          "amount"
        ],
        "type": "object",
        "properties": {
          "currency": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "PaymentDetails": {
        "type": "object",
        "properties": {
          "cardDetails": {
            "$ref": "#/components/schemas/CardDetails"
          },
          "transactionDetails": {
            "$ref": "#/components/schemas/TransactionDetails"
          }
        },
        "additionalProperties": false
      },
      "ProcessPaymentResult": {
        "type": "object",
        "properties": {
          "paymentId": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "success": {
            "type": "boolean",
            "readOnly": true
          },
          "transactionId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "MaskedCardDetails": {
        "type": "object",
        "properties": {
          "maskedCardNumber": {
            "type": "string"
          },
          "cardholderName": {
            "type": "string"
          },
          "expires": {
            "type": "string",
            "format": "date-time"
          },
          "validFrom": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PaymentResponse": {
        "type": "object",
        "properties": {
          "successful": {
            "type": "boolean"
          },
          "transactionId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "PaymentRequestLog": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "maskedCardDetails": {
            "$ref": "#/components/schemas/MaskedCardDetails"
          },
          "transactionDetails": {
            "$ref": "#/components/schemas/TransactionDetails"
          },
          "paymentResponse": {
            "$ref": "#/components/schemas/PaymentResponse"
          }
        },
        "additionalProperties": false
      }
    }
  }
}